+++ ORBIT/phases/install/__init__.py
@@ -6,7 +6,10 @@ __maintainer__ = ["Jake Nunemaker", "Rob Hammond"]
 __email__ = ["jake.nunemaker@nrel.gov" "robert.hammond@nrel.gov"]
 
 from .install_phase import InstallPhase  # isort:skip
-from .oss_install import OffshoreSubstationInstallation
+from .oss_install import (
+    FloatingSubstationInstallation,
+    OffshoreSubstationInstallation,
+)
 from .cable_install import ArrayCableInstallation, ExportCableInstallation
 from .mooring_install import MooringSystemInstallation
 from .turbine_install import TurbineInstallation