+++ tests/phases/install/oss_install/test_oss_install.py
@@ -15,7 +15,10 @@ from ORBIT import ProjectManager
 from tests.data import test_weather
 from ORBIT.core.library import extract_library_specs
 from ORBIT.core.defaults import process_times as pt
-from ORBIT.phases.install import OffshoreSubstationInstallation
+from ORBIT.phases.install import (
+    FloatingSubstationInstallation,
+    OffshoreSubstationInstallation,
+)
 from ORBIT.core.exceptions import MissingComponent
 
 config_single = extract_library_specs("config", "oss_install")
@@ -26,8 +29,8 @@ config_multi["num_feeders"] = 2
 
 @pytest.mark.parametrize(
     "config",
-    (config_single, config_multi, config_floating),
-    ids=["single_feeder", "multi_feeder", "floating"],
+    (config_single, config_multi),
+    ids=["single_feeder", "multi_feeder"],
 )
 def test_simulation_setup(config):
 
@@ -41,10 +44,18 @@ def test_simulation_setup(config):
     assert len(sim.port.items) == 2 * sim.num_substations
 
 
+def test_floating_simulation_setup():
+
+    sim = FloatingSubstationInstallation(config_floating)
+    assert sim.config == config_floating
+    assert sim.env
+    assert sim.num_substations == sim.config["num_substations"]
+
+
 @pytest.mark.parametrize(
     "config",
-    (config_single, config_multi, config_floating),
-    ids=["single_feeder", "multi_feeder", "floating"],
+    (config_single, config_multi),
+    ids=["single_feeder", "multi_feeder"],
 )
 def test_vessel_initialization(config):
 
@@ -64,8 +75,8 @@ def test_vessel_initialization(config):
 
 @pytest.mark.parametrize(
     "config",
-    (config_single, config_multi, config_floating),
-    ids=["single_feeder", "multi_feeder", "floating"],
+    (config_single, config_multi),
+    ids=["single_feeder", "multi_feeder"],
 )
 @pytest.mark.parametrize(
     "weather", (None, test_weather), ids=["no_weather", "test_weather"]
@@ -87,6 +98,23 @@ def test_for_complete_logging(weather, config):
     assert ~df["cost"].isnull().any()
     _ = sim.agent_efficiencies
     _ = sim.detailed_output
+
+
+@pytest.mark.parametrize(
+    "weather", (None, test_weather), ids=["no_weather", "test_weather"]
+)
+def test_for_complete_logging_floating(weather):
+
+    sim = FloatingSubstationInstallation(config_floating, weather=weather)
+    sim.run()
+
+    df = pd.DataFrame(sim.env.actions)
+    df = df.assign(shift=(df["time"] - df["time"].shift(1)))
+
+    for vessel in df["agent"].unique():
+        _df = df[df["agent"] == vessel].copy()
+        _df = _df.assign(shift=(_df["time"] - _df["time"].shift(1)))
+        assert (_df["shift"] - _df["duration"]).fillna(0.0).abs().max() < 1e-9
 
 
 def test_kwargs():